ticket:3108 Add support for WP85 external GPIO in sysfs

+++ include/linux/gpio/driver.h
@@ -120,7 +120,7 @@ struct gpio_chip {
 #endif
 /*SWISTART*/
 #ifdef CONFIG_SIERRA_EXT_GPIO
-	unsigned mask;
+	u64 mask;
 #endif /*CONFIG_SIERRA_EXT_GPIO*/
 /*SWISTOP*/
 };
